requirements.txt
django>=2.0,<3.0
djangorestframework
psycopg2>=2.7,<3.0
django-adminlte3
docxtpl
